+++ package.json
@@ -11,6 +11,8 @@
     "update:smt": "npx mocha ./test/merkle-tree/**.js --update",
     "test:st:no-data": "npx mocha ./test/state-transition/no-data/st-all.test.js",
     "test:st:calldata": "npx mocha ./test/state-transition/calldata/st-all.test.js",
+    "test:mt:bridge": "npx mocha ./test/mt-bridge/**.test.js",
+    "update:mt:bridge": "npx mocha ./test/mt-bridge/**.test.js --update",
     "update:st:calldata": "cd ./tools-calldata/evm && ./gen-all.sh",
     "update:st:no-data": "npx mocha ./test/state-transition/no-data/st-all.test.js --update",
     "test:contracts": "npx mocha ./test/state-transition/no-data/contracts.test.js"
